Make it work without animation
