+++ src/app/item-popup/ItemPerksList.m.scss.d.ts
@@ -1,15 +1,13 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'archetype': string;
-  'archetypeMod': string;
   'disabled': string;
-  'mod': string;
   'perkIcon': string;
+  'perkInfo': string;
   'plug': string;
   'plugLabel': string;
   'plugged': string;
-  'row': string;
+  'selected': string;
   'socket': string;
   'sockets': string;
 }
