+++ src/app/item-popup/ItemSocketsWeapons.tsx
@@ -15,6 +15,7 @@ import { D2ManifestDefinitions } from '../destiny2/d2-definitions';
 import { DimItem, DimPlug, DimSocket } from '../inventory/item-types';
 import { inventoryWishListsSelector, wishListsEnabledSelector } from '../wishlists/selectors';
 import { InventoryWishListRoll } from '../wishlists/wishlists';
+import ItemPerksList from './ItemPerksList';
 import './ItemSockets.scss';
 import styles from './ItemSocketsWeapons.m.scss';
 import Plug from './Plug';
@@ -170,36 +171,39 @@ function ItemSocketsWeapons({
           )}
         </div>
       )}
-      {perks && (
-        <div
-          className={clsx(
-            'item-socket-category',
-            categoryStyle(perks.category.categoryStyle),
-            styles.perks
-          )}
-        >
-          <div className="item-sockets">
-            {perks.sockets.map(
-              (socketInfo) =>
-                socketInfo.socketDefinition.socketTypeHash !== killTrackerSocketTypeHash && (
-                  <Socket
-                    key={socketInfo.socketIndex}
-                    defs={defs}
-                    item={item}
-                    isPhonePortrait={isPhonePortrait}
-                    socket={socketInfo}
-                    wishListsEnabled={wishListsEnabled}
-                    inventoryWishListRoll={inventoryWishListRoll}
-                    classesByHash={classesByHash}
-                    onClick={handleSocketClick}
-                    onShiftClick={onShiftClick}
-                    adjustedPlug={adjustedItemPlugs?.[socketInfo.socketIndex]}
-                  />
-                )
+      {perks &&
+        ($featureFlags.newPerks ? (
+          <ItemPerksList item={item} perks={perks} />
+        ) : (
+          <div
+            className={clsx(
+              'item-socket-category',
+              categoryStyle(perks.category.categoryStyle),
+              styles.perks
             )}
+          >
+            <div className="item-sockets">
+              {perks.sockets.map(
+                (socketInfo) =>
+                  socketInfo.socketDefinition.socketTypeHash !== killTrackerSocketTypeHash && (
+                    <Socket
+                      key={socketInfo.socketIndex}
+                      defs={defs}
+                      item={item}
+                      isPhonePortrait={isPhonePortrait}
+                      socket={socketInfo}
+                      wishListsEnabled={wishListsEnabled}
+                      inventoryWishListRoll={inventoryWishListRoll}
+                      classesByHash={classesByHash}
+                      onClick={handleSocketClick}
+                      onShiftClick={onShiftClick}
+                      adjustedPlug={adjustedItemPlugs?.[socketInfo.socketIndex]}
+                    />
+                  )
+              )}
+            </div>
           </div>
-        </div>
-      )}
+        ))}
       {minimal && mods.length > 0 && (
         <div className="item-socket-category-Consumable socket-container">
           {mods.map((socketInfo) => (
